Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    int a,b,s;
+    float a,b,s;
     char ch;
     scanf("%d%d%c",&a,&b, &ch);
     if(ch=='+'){
